rename hopson_bot.js to main.js
