+++ package.json
@@ -5,7 +5,7 @@
   "main": "scripts/bot.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node src/hopson_bot.js"
+    "start": "node src/main.js"
   },
   "repository": {
     "type": "git",
