+++ src/quiz/quiz_command_handler.js
@@ -3,7 +3,7 @@ const QuizJSON      = require("../../" + questionsFile);
 
 const CommandHandlerBase = require("../command_handler_base")
 const Config        = require('../../data/config.json');
-const Bot           = require("../hopson_bot");
+const Bot           = require("../main");
 const Util          = require("../misc/util");
 const QuizSession   = require("./quiz_session")
