+++ src/quiz/quiz_session.js
@@ -1,6 +1,6 @@
 const questionsFile = "data/quiz_questions.json";
 
-const Bot           = require("../hopson_bot");
+const Bot           = require("../main");
 const Util          = require("../misc/util");
 const JSONFile      = require('jsonfile');
 const Discord       = require('discord.js')
